% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerPredictiveBiomarker.shiny.R
\name{PowerPredictiveBiomarker.shiny}
\alias{PowerPredictiveBiomarker.shiny}
\title{Run for Shiny Applictaion for Power Calculation of Predictive Biomarker and generate a statistical plan to justify the sample size}
\usage{
PowerPredictiveBiomarker.shiny()
}
\description{
Run for Shiny Applictaion for Power Calculation of Predictive Biomarker and generate a statistical plan to justify the sample size
}

